Refactor comparison into spanalyzer
