+++ tracing-test-utils/src/main/java/com/palantir/tracing/TestTracingExtension.java
@@ -26,13 +26,9 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Collection;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Set;
-import java.util.function.Function;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-import java.util.stream.Stream;
 import org.junit.jupiter.api.extension.AfterEachCallback;
 import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
@@ -97,7 +93,7 @@ final class TestTracingExtension implements BeforeEachCallback, AfterEachCallbac
         SpanAnalyzer.Result expected = SpanAnalyzer.analyze(expectedSpans);
         SpanAnalyzer.Result actual = SpanAnalyzer.analyze(actualSpans);
 
-        Set<ComparisonFailure> failures = compareSpansRecursively(expected, actual, expected.root(), actual.root())
+        Set<ComparisonFailure> failures = SpanAnalyzer.compareSpansRecursively(expected, actual, expected.root(), actual.root())
                 .collect(ImmutableSet.toImmutableSet());
 
         HtmlFormatter.render(HtmlFormatter.RenderConfig.builder()
@@ -147,95 +143,6 @@ final class TestTracingExtension implements BeforeEachCallback, AfterEachCallbac
         }
     }
 
-    private static Stream<ComparisonFailure> compareSpansRecursively(
-            SpanAnalyzer.Result expected,
-            SpanAnalyzer.Result actual,
-            Span ex,
-            Span ac) {
-        if (!ex.getOperation().equals(ac.getOperation())) {
-            return Stream.of(ComparisonFailure.unequalOperation(ex, ac));
-        }
-        // other fields, type, params, metadata(???)
-
-        // ensure we have the same number of children, same child operation names in the same order
-        List<Span> sortedExpectedChildren = SpanAnalyzer.children(expected.graph(), ex)
-                .sorted(Comparator.comparingLong(Span::getStartTimeMicroSeconds))
-                .collect(ImmutableList.toImmutableList());
-        List<Span> sortedActualChildren = SpanAnalyzer.children(actual.graph(), ac)
-                .sorted(Comparator.comparingLong(Span::getStartTimeMicroSeconds))
-                .collect(ImmutableList.toImmutableList());
-
-        if (sortedExpectedChildren.size() != sortedActualChildren.size()) {
-            // just highlighting the parents for now.
-            return Stream.of(ComparisonFailure.unequalChildren(ex, ac, sortedExpectedChildren, sortedActualChildren));
-        }
-
-        boolean expectedContainsOverlappingSpans = containsOverlappingSpans(sortedExpectedChildren);
-        boolean actualContainsOverlappingSpans = containsOverlappingSpans(sortedActualChildren);
-        if (expectedContainsOverlappingSpans ^ actualContainsOverlappingSpans) {
-            // Either Both or neither tree should have concurrent spans
-            return Stream.of(ComparisonFailure.incompatibleStructure(ex, ac));
-        }
-
-        if (actualContainsOverlappingSpans && !compatibleOverlappingSpans(
-                expected, actual, sortedExpectedChildren, sortedActualChildren)) {
-            return Stream.of(ComparisonFailure.unequalChildren(ex, ac, sortedExpectedChildren, sortedActualChildren));
-        }
-
-        return IntStream.range(0, sortedActualChildren.size())
-                .mapToObj(i -> compareSpansRecursively(
-                        expected,
-                        actual,
-                        sortedExpectedChildren.get(i),
-                        sortedActualChildren.get(i)))
-                .flatMap(Function.identity());
-    }
-
-    /**
-     * When async spans are involved, there can be many overlapping children with the same operation name.
-     * We exhaustively check each possible pair, and require that each span in the 'expected' list lines up with
-     * something and each span in the 'actual' list also lines up with something.
-     *
-     * It's OK for some spans to be compatible with more than one span (as subtrees could be identical).
-     */
-    private static boolean compatibleOverlappingSpans(
-            SpanAnalyzer.Result expected, SpanAnalyzer.Result actual, List<Span> ex, List<Span> ac) {
-        boolean[][] compatibility = new boolean[ex.size()][ac.size()];
-
-        for (int exIndex = 0; exIndex < ex.size(); exIndex++) {
-            for (int acIndex = 0; acIndex < ac.size(); acIndex++) {
-                long numFailures = compareSpansRecursively(expected, actual, ex.get(exIndex), ac.get(acIndex)).count();
-                compatibility[exIndex][acIndex] = numFailures == 0;
-            }
-        }
-
-        // check rows first
-        for (int exIndex = 0; exIndex < ex.size(); exIndex++) {
-            boolean atLeastOneCompatible = false;
-            for (int acIndex = 0; acIndex < ac.size(); acIndex++) {
-                atLeastOneCompatible |= compatibility[exIndex][acIndex];
-            }
-
-            if (!atLeastOneCompatible) {
-                return false;
-            }
-        }
-
-        // check columns
-        for (int acIndex = 0; acIndex < ac.size(); acIndex++) {
-            boolean atLeastOneCompatible = false;
-            for (int exIndex = 0; exIndex < ex.size(); exIndex++) {
-                atLeastOneCompatible |= compatibility[exIndex][acIndex];
-            }
-
-            if (!atLeastOneCompatible) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
     private static String renderFailure(ComparisonFailure failure) {
         return failure.map(
                 (ComparisonFailure.unequalOperation t) -> String.format("Expected operation %s but received %s",
@@ -248,36 +155,8 @@ final class TestTracingExtension implements BeforeEachCallback, AfterEachCallbac
                         "Expected children to structured similarly"));
     }
 
-    /* Assumes list of spans to be ordered by startTimeMicros */
-    private static boolean containsOverlappingSpans(List<Span> spans) {
-        if (!spans.isEmpty()) {
-            Span currentSpan = spans.get(0);
-            for (int i = 1; i < spans.size(); i++) {
-                Span nextSpan = spans.get(i);
-                if (nextSpan.getStartTimeMicroSeconds() < getEndTimeMicroSeconds(currentSpan)) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-
-    private static long getEndTimeMicroSeconds(Span span) {
-        return span.getStartTimeMicroSeconds() + (span.getDurationNanoSeconds() * 1000);
-    }
-
     private static String testName(ExtensionContext context) {
         return context.getRequiredTestClass().getSimpleName() + "/" + context.getRequiredTestMethod().getName();
     }
 
-    @SuppressWarnings("checkstyle:TypeName")
-    @DataEnum
-    interface ComparisonFailure_dataenum {
-        dataenum_case unequalOperation(Span expected, Span actual);
-
-        dataenum_case unequalChildren(
-                Span expected, Span actual, List<Span> expectedChildren, List<Span> actualChildren);
-
-        dataenum_case incompatibleStructure(Span expected, Span actual);
-    }
 }
